refactor: separate logic from UI
Moved mutation logic and handlers to `Deployment` page, simplified
`DeploymentDetails`, and centralized shared error handling to reduce
duplication.
