+++ frontend/src/i18n/langs/en.json
@@ -470,9 +470,6 @@
   "components.DeploymentDetails.containerWithIndex": {
     "defaultMessage": "Container {index}"
   },
-  "components.DeploymentDetails.deleteErrorOffline": {
-    "defaultMessage": "The device is disconnected. You cannot delete an application while it is offline."
-  },
   "components.DeploymentDetails.deleteModal.description": {
     "defaultMessage": "This action cannot be undone. This will permanently delete the deployment."
   },
@@ -481,9 +478,6 @@
   },
   "components.DeploymentDetails.deleteModal.title": {
     "defaultMessage": "Delete Deployment"
-  },
-  "components.DeploymentDetails.deletionErrorFeedback": {
-    "defaultMessage": "Could not delete the deployment, please try again."
   },
   "components.DeploymentDetails.deviceMappings": {
     "defaultMessage": "Device Mappings"
@@ -512,35 +506,11 @@
   "components.DeploymentDetails.selectOption": {
     "defaultMessage": "Select a Release Version"
   },
-  "components.DeploymentDetails.sendErrorFeedback": {
-    "defaultMessage": "Could not send the Application to the device, please try again."
-  },
-  "components.DeploymentDetails.sendErrorOffline": {
-    "defaultMessage": "The device is disconnected. You cannot deploy an application while it is offline."
-  },
-  "components.DeploymentDetails.startErrorFeedback": {
-    "defaultMessage": "Could not Start the Deployed Application, please try again."
-  },
-  "components.DeploymentDetails.startErrorOffline": {
-    "defaultMessage": "The device is disconnected. You cannot start an application while it is offline."
-  },
-  "components.DeploymentDetails.stopErrorFeedback": {
-    "defaultMessage": "Could not Stop the Deployed Application, please try again."
-  },
-  "components.DeploymentDetails.stopErrorOffline": {
-    "defaultMessage": "The device is disconnected. You cannot stop an application while it is offline."
-  },
   "components.DeploymentDetails.unnamed": {
     "defaultMessage": "Unnamed"
   },
   "components.DeploymentDetails.unnamedImage": {
     "defaultMessage": "Unnamed image"
-  },
-  "components.DeploymentDetails.upgradeErrorFeedback": {
-    "defaultMessage": "Could not upgrade the deployment, please try again."
-  },
-  "components.DeploymentDetails.upgradeErrorOffline": {
-    "defaultMessage": "The device is disconnected. You cannot upgrade an application while it is offline."
   },
   "components.DeploymentDetails.volumes": {
     "defaultMessage": "Volumes"
@@ -3175,11 +3145,41 @@
   "pages.Channels.title": {
     "defaultMessage": "Channels"
   },
+  "pages.Deployment.deleteErrorOffline": {
+    "defaultMessage": "The device is disconnected. You cannot delete an application while it is offline."
+  },
+  "pages.Deployment.deletionErrorFeedback": {
+    "defaultMessage": "Could not delete the deployment, please try again."
+  },
   "pages.Deployment.deploymentNotFound.message": {
     "defaultMessage": "Return to the device page."
   },
   "pages.Deployment.deploymentNotFound.title": {
     "defaultMessage": "Deployment not found."
+  },
+  "pages.Deployment.sendErrorFeedback": {
+    "defaultMessage": "Could not send the Application to the device, please try again."
+  },
+  "pages.Deployment.sendErrorOffline": {
+    "defaultMessage": "The device is disconnected. You cannot deploy an application while it is offline."
+  },
+  "pages.Deployment.startErrorFeedback": {
+    "defaultMessage": "Could not Start the Deployed Application, please try again."
+  },
+  "pages.Deployment.startErrorOffline": {
+    "defaultMessage": "The device is disconnected. You cannot start an application while it is offline."
+  },
+  "pages.Deployment.stopErrorFeedback": {
+    "defaultMessage": "Could not Stop the Deployed Application, please try again."
+  },
+  "pages.Deployment.stopErrorOffline": {
+    "defaultMessage": "The device is disconnected. You cannot stop an application while it is offline."
+  },
+  "pages.Deployment.upgradeErrorFeedback": {
+    "defaultMessage": "Could not upgrade the deployment, please try again."
+  },
+  "pages.Deployment.upgradeErrorOffline": {
+    "defaultMessage": "The device is disconnected. You cannot upgrade an application while it is offline."
   },
   "pages.DeploymentCampaign.deploymentCampaignNotFound.message": {
     "defaultMessage": "Return to the Deployment Campaign list."
