+++ frontend/src/pages/Deployment.tsx
@@ -26,6 +26,7 @@ import { FormattedMessage } from "react-intl";
 import type { PreloadedQuery } from "react-relay/hooks";
 import {
   graphql,
+  useMutation,
   usePreloadedQuery,
   useQueryLoader,
   useSubscription,
@@ -37,14 +38,39 @@ import type {
   Deployment_getDeployment_Query$data,
 } from "@/api/__generated__/Deployment_getDeployment_Query.graphql";
 import type { Deployment_deployment_updated_Subscription } from "@/api/__generated__/Deployment_deployment_updated_Subscription.graphql";
+import type { Deployment_sendDeployment_Mutation } from "@/api/__generated__/Deployment_sendDeployment_Mutation.graphql";
+import type { Deployment_startDeployment_Mutation } from "@/api/__generated__/Deployment_startDeployment_Mutation.graphql";
+import type { Deployment_stopDeployment_Mutation } from "@/api/__generated__/Deployment_stopDeployment_Mutation.graphql";
+import type { Deployment_upgradeDeployment_Mutation } from "@/api/__generated__/Deployment_upgradeDeployment_Mutation.graphql";
+import type { Deployment_deleteDeployment_Mutation } from "@/api/__generated__/Deployment_deleteDeployment_Mutation.graphql";
 
-import { Link, Route } from "@/Navigation";
+import { Link, Route, useNavigate } from "@/Navigation";
 import Alert from "@/components/Alert";
 import Center from "@/components/Center";
 import DeploymentDetails from "@/components/DeploymentDetails";
 import Page from "@/components/Page";
 import Result from "@/components/Result";
 import Spinner from "@/components/Spinner";
+
+const handleMutationCompletion = (
+  errors: readonly any[] | null | undefined,
+  setErrorFeedback: (msg: React.ReactNode) => void,
+  onSuccess?: () => void,
+) => {
+  if (errors && errors.length > 0) {
+    const errorMessage = errors
+      .map(({ fields, message }) =>
+        fields?.length ? `${fields.join(" ")} ${message}` : message,
+      )
+      .join(". \n");
+
+    setErrorFeedback(errorMessage);
+    return;
+  }
+
+  setErrorFeedback(null);
+  onSuccess?.();
+};
 
 const GET_DEPLOYMENT_QUERY = graphql`
   query Deployment_getDeployment_Query(
@@ -129,6 +155,69 @@ const DEPLOYMENT_UPDATED_SUBSCRIPTION = graphql`
   }
 `;
 
+const SEND_DEPLOYMENT_MUTATION = graphql`
+  mutation Deployment_sendDeployment_Mutation($id: ID!) {
+    sendDeployment(id: $id) {
+      result {
+        id
+        state
+      }
+      errors {
+        message
+      }
+    }
+  }
+`;
+
+const START_DEPLOYMENT_MUTATION = graphql`
+  mutation Deployment_startDeployment_Mutation($id: ID!) {
+    startDeployment(id: $id) {
+      result {
+        id
+      }
+      errors {
+        message
+      }
+    }
+  }
+`;
+
+const STOP_DEPLOYMENT_MUTATION = graphql`
+  mutation Deployment_stopDeployment_Mutation($id: ID!) {
+    stopDeployment(id: $id) {
+      result {
+        id
+      }
+      errors {
+        message
+      }
+    }
+  }
+`;
+
+const DELETE_DEPLOYMENT_MUTATION = graphql`
+  mutation Deployment_deleteDeployment_Mutation($id: ID!) {
+    deleteDeployment(id: $id) {
+      result {
+        id
+      }
+    }
+  }
+`;
+
+const UPGRADE_DEPLOYMENT_MUTATION = graphql`
+  mutation Deployment_upgradeDeployment_Mutation(
+    $id: ID!
+    $input: UpgradeDeploymentInput!
+  ) {
+    upgradeDeployment(id: $id, input: $input) {
+      result {
+        id
+      }
+    }
+  }
+`;
+
 interface DeploymentContentProps {
   deployment: NonNullable<Deployment_getDeployment_Query$data["deployment"]>;
   isOnline: boolean;
@@ -138,7 +227,195 @@ const DeploymentContent = ({
   deployment,
   isOnline,
 }: DeploymentContentProps) => {
+  const navigate = useNavigate();
+  const { deviceId } = useParams();
+
   const [errorFeedback, setErrorFeedback] = useState<React.ReactNode>(null);
+
+  const [sendDeployment] = useMutation<Deployment_sendDeployment_Mutation>(
+    SEND_DEPLOYMENT_MUTATION,
+  );
+  const [startDeployment] = useMutation<Deployment_startDeployment_Mutation>(
+    START_DEPLOYMENT_MUTATION,
+  );
+  const [stopDeployment] = useMutation<Deployment_stopDeployment_Mutation>(
+    STOP_DEPLOYMENT_MUTATION,
+  );
+
+  const [deleteDeployment, isDeletingDeployment] =
+    useMutation<Deployment_deleteDeployment_Mutation>(
+      DELETE_DEPLOYMENT_MUTATION,
+    );
+
+  const [upgradeDeployment] =
+    useMutation<Deployment_upgradeDeployment_Mutation>(
+      UPGRADE_DEPLOYMENT_MUTATION,
+    );
+
+  const handleSendDeployedApplication = useCallback(
+    (deploymentId: string) => {
+      if (!isOnline) {
+        return setErrorFeedback(
+          <FormattedMessage
+            id="pages.Deployment.sendErrorOffline"
+            defaultMessage="The device is disconnected. You cannot deploy an application while it is offline."
+          />,
+        );
+      }
+
+      sendDeployment({
+        variables: { id: deploymentId },
+        onCompleted: (_data, errors) => {
+          handleMutationCompletion(errors, setErrorFeedback);
+        },
+        onError: () => {
+          setErrorFeedback(
+            <FormattedMessage
+              id="pages.Deployment.sendErrorFeedback"
+              defaultMessage="Could not send the Application to the device, please try again."
+            />,
+          );
+        },
+      });
+    },
+    [isOnline, sendDeployment, setErrorFeedback],
+  );
+
+  const handleStartDeployedApplication = useCallback(
+    (deploymentId: string) => {
+      if (!isOnline) {
+        return setErrorFeedback(
+          <FormattedMessage
+            id="pages.Deployment.startErrorOffline"
+            defaultMessage="The device is disconnected. You cannot start an application while it is offline."
+          />,
+        );
+      }
+
+      startDeployment({
+        variables: { id: deploymentId },
+        onCompleted: (_data, errors) => {
+          handleMutationCompletion(errors, setErrorFeedback);
+        },
+        onError: () => {
+          setErrorFeedback(
+            <FormattedMessage
+              id="pages.Deployment.startErrorFeedback"
+              defaultMessage="Could not Start the Deployed Application, please try again."
+            />,
+          );
+        },
+      });
+    },
+    [isOnline, startDeployment, setErrorFeedback],
+  );
+
+  const handleStopDeployedApplication = useCallback(
+    (deploymentId: string) => {
+      if (!isOnline) {
+        return setErrorFeedback(
+          <FormattedMessage
+            id="pages.Deployment.stopErrorOffline"
+            defaultMessage="The device is disconnected. You cannot stop an application while it is offline."
+          />,
+        );
+      }
+
+      stopDeployment({
+        variables: { id: deploymentId },
+        onCompleted: (_data, errors) => {
+          handleMutationCompletion(errors, setErrorFeedback);
+        },
+        onError: () => {
+          setErrorFeedback(
+            <FormattedMessage
+              id="pages.Deployment.stopErrorFeedback"
+              defaultMessage="Could not Stop the Deployed Application, please try again."
+            />,
+          );
+        },
+      });
+    },
+    [isOnline, stopDeployment, setErrorFeedback],
+  );
+
+  const handleDeleteDeployedApplication = useCallback(
+    (deploymentId: string) => {
+      if (!isOnline) {
+        return setErrorFeedback(
+          <FormattedMessage
+            id="pages.Deployment.deleteErrorOffline"
+            defaultMessage="The device is disconnected. You cannot delete an application while it is offline."
+          />,
+        );
+      }
+
+      deleteDeployment({
+        variables: { id: deploymentId },
+        onCompleted: (_data, errors) => {
+          handleMutationCompletion(errors, setErrorFeedback, () => {
+            if (deviceId) {
+              navigate({
+                route: Route.devicesEdit,
+                params: { deviceId },
+              });
+            }
+          });
+        },
+        onError: () => {
+          setErrorFeedback(
+            <FormattedMessage
+              id="pages.Deployment.deletionErrorFeedback"
+              defaultMessage="Could not delete the deployment, please try again."
+            />,
+          );
+        },
+      });
+    },
+    [deleteDeployment, setErrorFeedback, navigate, deviceId, isOnline],
+  );
+
+  const handleUpgradeDeployedRelease = useCallback(
+    (deploymentId: string, upgradeTargetReleaseId: string) => {
+      if (!isOnline) {
+        return setErrorFeedback(
+          <FormattedMessage
+            id="pages.Deployment.upgradeErrorOffline"
+            defaultMessage="The device is disconnected. You cannot upgrade an application while it is offline."
+          />,
+        );
+      }
+
+      upgradeDeployment({
+        variables: {
+          id: deploymentId,
+          input: { target: upgradeTargetReleaseId },
+        },
+        onCompleted: (data, errors) => {
+          handleMutationCompletion(errors, setErrorFeedback, () => {
+            if (deviceId && data?.upgradeDeployment?.result?.id) {
+              navigate({
+                route: Route.deploymentEdit,
+                params: {
+                  deviceId,
+                  deploymentId: data.upgradeDeployment.result.id,
+                },
+              });
+            }
+          });
+        },
+        onError() {
+          setErrorFeedback(
+            <FormattedMessage
+              id="pages.Deployment.upgradeErrorFeedback"
+              defaultMessage="Could not upgrade the deployment, please try again."
+            />,
+          );
+        },
+      });
+    },
+    [upgradeDeployment, setErrorFeedback, isOnline, deviceId, navigate],
+  );
 
   return (
     <Page className="h-100 d-flex flex-column overflow-hidden">
@@ -162,8 +439,13 @@ const DeploymentContent = ({
 
         <DeploymentDetails
           deploymentRef={deployment}
-          isOnline={isOnline}
+          isDeletingDeployment={isDeletingDeployment}
           setErrorFeedback={setErrorFeedback}
+          onStart={handleStartDeployedApplication}
+          onStop={handleStopDeployedApplication}
+          onRedeploy={handleSendDeployedApplication}
+          onDelete={handleDeleteDeployedApplication}
+          onUpgrade={handleUpgradeDeployedRelease}
         />
       </Page.Main>
     </Page>
@@ -234,14 +516,14 @@ const DeploymentPage = () => {
   return (
     <Suspense
       fallback={
-        <Center data-testid="page-loading">
+        <Center>
           <Spinner />
         </Center>
       }
     >
       <ErrorBoundary
         FallbackComponent={(props) => (
-          <Center data-testid="page-error">
+          <Center>
             <Page.LoadingError onRetry={props.resetErrorBoundary} />
           </Center>
         )}
